Combine AI providers into 1 config file && move provider fetching code to this config file
