--- chatapi/src/config/gemini.config.ts
@@ -1,9 +1,0 @@
-import { GoogleGenerativeAI } from '@google/generative-ai';
-import dotenv from 'dotenv';
-
-dotenv.config();
-
-const gemini = new GoogleGenerativeAI(process.env.GEMINI_API_KEY || '');
-
-export default gemini;
-
