--- chatapi/src/config/openai.config.ts
@@ -1,10 +1,0 @@
-import OpenAI from 'openai';
-import dotenv from 'dotenv';
-
-dotenv.config();
-
-const openai = new OpenAI({
-  'apiKey': process.env.OPENAI_API_KEY || '',
-});
-
-export default openai;
